Trim leading and trailing spaces in CR server IP address
Under the "Cache Refresh" form, a user is able to insert the oxTrust server
IP. As per documentation, even in clustered mode, this will be a single
of the server that will fetch data from LDAP/AD instances. An issue
arose due to lack of validation that the user inserts a valid IP address.

As a first step, in order to provide a better UX and avoid annoying
issues where spaces find their way into the form field (either pasted or
by mistake), all spaces should be trimmed on the backend.

Signed-off-by: Linos Giannopoulos <14967949+linosgian@users.noreply.github.com>